+++ src/app/routes/onDemandRadio/getInitialData/index.js
@@ -1,7 +1,7 @@
 import path from 'ramda/src/path';
 import fetchPageData from '../../utils/fetchPageData';
 import overrideRendererOnTest from '../../utils/overrideRendererOnTest';
-import { getPlaceholderImageUrl } from '../../utils/imageUrl';
+import getPlaceholderImageUrl from '../../utils/getPlaceholderImageUrl';
 
 const getBrandTitle = path(['metadata', 'title']);
 const getLanguage = path(['metadata', 'language']);
